Rename WarehouseImport extension to ProductsImport
