+++ hugashop/crm/Extensions/ProductsImport/settings.yaml
@@ -1,3 +1,3 @@
-name: "Импорт в склад"
+name: "Импорт товаров на склад"
 description: "Загрузка CSV-файла с товарами и добавление их на выбранный склад"
-version: 1.1
+version: 1.2
